Fort Collins weather data

This data is extracted from the extRemes package and consists of daily cumulated precipitation (in mm), minimum and maximum temperature (in degree Celcius). The astute reader will notice that the precipitation series (reported in inches according to the documentation) has been multiplied by 100.

FortCollins

Format

A data frame with 36524 rows and four variables:

date

Date of observation

precip

daily cumulated precipitation (in mm); most are zero

maxTemp

maximum daily temperature temperature (in Celcius)

minTemp

minimum daily temperature (in Celcius)

Source

Colorado Climate Center at Colorado State University

References

Katz, R. W., Parlange, M. B. and Naveau, P. (2002) Statistics of extremes in hydrology. Advances in Water Resources, 25, 1287–1304.

See also