Daily returns for three major US banks

This data frame contains daily returns for three major banks, namely Chase (C), Bank of America (BAC) and Goldman Sachs (GS) for the period 2000-2020. The data were extracted using `tidyquant` and computed from adjusted returns, with the difference relative to the previous day of operation, with \(r_t=(Y_t-Y_{t-1})/Y_{t-1}\). The date field contains the date during which the return is calculated.

returnsBanks

Format

A data frame with 5283 rows and the following four variables

date

Date at which the return is calculated

retC

returns for Chase

retBAC

returns for Bank of America

retGS

returns for Goldman Sachs